[refactor] add hand to player
