+++ lib/classes/player/dealer.rb
@@ -2,4 +2,8 @@ class Dealer < Player
   def initialize(points_counting_rule)
     super('Dealer', points_counting_rule)
   end
+
+  def should_hit?
+    points < 17 && !full?
+  end
 end
